Update program 3.cpp Round Robin
DESCRIPTION:
To aim is to calculate the average waiting time. There will be a time slice, each 
process should be executed within that time-slice and if not it will go to the waiting 
state so first check whether the burst time is less than the time-slice. If it is less than it 
assign the waiting time to the sum of the total times. If it is greater than the burst-time 
then subtract the time slot from the actual burst time and increment it by time-slot and 
the loop continues until all the processes are completed.