Update Tower of hanoi.c
The purpose of this code is to implement the Tower of Hanoi problem, a classic problem in recursion and algorithmic thinking. The Tower of Hanoi consists of three rods and a number of disks of different sizes, which can slide onto any rod. The puzzle starts with the disks in a neat stack in ascending order of size on one rod, the smallest at the top, thus making a conical shape.